+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - Update Agency API ValidationError
+- Return device_id for 404 when device doesn't not exist
 
 
 0.6.0 (2019-09-02)